refactor(upper_assembly): restructure UpperAssembly hierarchy
- Moved `UpperAssembly` interface to a new package `upper_assembly`.
- Introduced `UpperAssemblyBase` abstract class extending `SubsystemBase`.
- Updated implementations (`NarwhalUpperAssembly`, `SquidUpperAssembly`) to extend `UpperAssemblyBase`.
- Adjusted imports and references across the codebase to reflect these changes.
