+++ src/main/java/frc/robot/subsystems/squid/SquidUpperAssembly.java
@@ -8,10 +8,10 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.RunCommand;
 import frc.robot.commands.squid.ManualSquidManipulatorCommand;
-import frc.robot.subsystems.UpperAssembly;
+import frc.robot.subsystems.upper_assembly.UpperAssemblyBase;
 import frc.robot.util.ScoringHeight;
 
-public class SquidUpperAssembly implements UpperAssembly {
+public class SquidUpperAssembly extends UpperAssemblyBase {
 
     SquidManipulator squidManipulator;
