+++ src/main/java/frc/robot/subsystems/upper_assembly/UpperAssembly.java
@@ -1,11 +1,10 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.upper_assembly;
 
 import java.util.function.Supplier;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.Subsystem;
 import frc.robot.util.ScoringHeight;
 
 /**
@@ -13,7 +12,7 @@ import frc.robot.util.ScoringHeight;
  * Provides methods for various assembly-specific actions such as scoring, intaking, climbing, 
  * and manual control. 
  */
-public interface UpperAssembly extends Subsystem {
+public interface UpperAssembly {
 
     /**
      * Creates a command to run the coral intake mechanism.
